+++ base/components/AutoCompounder/Modals/DepositConfirmationModal.tsx
@@ -1,28 +1,29 @@
-import { RemoveRedEye } from "@mui/icons-material";
-import {
-  Box,
-  CircularProgress,
-  IconButton,
-  InputLabel,
-  Stack,
-  Tooltip,
-  Typography,
-} from "@mui/material";
+import CloseIcon from "@mui/icons-material/Close";
+import { Box, CircularProgress, Stack, Typography } from "@mui/material";
 import { ERC20_ABI } from "abis/ERC20";
 import { VAULT_ABI } from "abis/Vault";
 import { parseUnits } from "ethers";
 import { LiquidityPool } from "lib/types";
-import React, { useMemo, useState } from "react";
-import { toast } from "react-toastify";
+import React, { useEffect, useMemo, useState } from "react";
+import { Address } from "viem";
 import {
-  Address,
   useAccount,
   useContractRead,
   useContractWrite,
   useWaitForTransaction,
 } from "wagmi";
 import BaseModal from "./BaseModal";
-import { ActionButton, InputField, StepButton, StepContainer } from "./styles";
+import {
+  ActionButton,
+  CloseButton,
+  HeaderTitle,
+  InputField,
+  InputLabel,
+  ModalContent,
+  ModalHeader,
+  StepButton,
+  StepContainer,
+} from "./styles";
 
 interface DepositConfirmationModalProps {
   open: boolean;
@@ -32,6 +33,14 @@ interface DepositConfirmationModalProps {
   onSuccess?: () => void;
 }
 
+type TransactionState =
+  | "idle"
+  | "approving"
+  | "approved"
+  | "depositing"
+  | "success"
+  | "error";
+
 const DepositConfirmationModal: React.FC<DepositConfirmationModalProps> = ({
   open,
   onClose,
@@ -39,8 +48,8 @@ const DepositConfirmationModal: React.FC<DepositConfirmationModalProps> = ({
   lpToken,
   onSuccess,
 }) => {
-  const [transactionToastId, setTransactionToastId] =
-    useState<React.ReactText | null>(null);
+  const [transactionState, setTransactionState] =
+    useState<TransactionState>("idle");
   const { address } = useAccount();
 
   const amountInWei = useMemo(() => {
@@ -52,8 +61,7 @@ const DepositConfirmationModal: React.FC<DepositConfirmationModalProps> = ({
     }
   }, [amount, lpToken.decimals]);
 
-  // Read allowance
-  const { data: allowance } = useContractRead({
+  const { data: allowance = BigInt(0) } = useContractRead({
     address: lpToken.poolAddress,
     abi: ERC20_ABI,
     functionName: "allowance",
@@ -62,328 +70,242 @@ const DepositConfirmationModal: React.FC<DepositConfirmationModalProps> = ({
     watch: true,
   });
 
-  // Approve transaction
-  const {
-    write: approve,
-    data: approveData,
-    isLoading: isApproveLoading,
-    reset: resetApprove,
-  } = useContractWrite({
+  const isApprovalNeeded = allowance < amountInWei;
+
+  // Reset state when modal opens/closes
+  useEffect(() => {
+    if (open) {
+      setTransactionState(isApprovalNeeded ? "idle" : "approved");
+    } else {
+      setTransactionState("idle");
+    }
+  }, [open, isApprovalNeeded]);
+
+  const { write: approve, data: approveData } = useContractWrite({
     address: lpToken.poolAddress,
     abi: ERC20_ABI,
     functionName: "approve",
-    onError: (error) => handleTransactionError(error, "approve", "pre"),
   });
 
-  // Watch approve transaction
-  const { isLoading: isApproveConfirming, isSuccess: isApproveSuccess } =
-    useWaitForTransaction({
-      hash: approveData?.hash,
-      confirmations: 2,
-      onSuccess: () => {
-        handleApproveSuccess();
-      },
-      onError: (error) => handleTransactionError(error, "approve", "onChain"),
-    });
-
-  // Deposit transaction
-  const {
-    write: deposit,
-    data: depositData,
-    isLoading: isDepositLoading,
-    reset: resetDeposit,
-  } = useContractWrite({
+  const { write: deposit, data: depositData } = useContractWrite({
     address: lpToken.vault,
     abi: VAULT_ABI,
     functionName: "deposit",
-    onError: (error) => handleTransactionError(error, "deposit", "pre"),
   });
 
-  // Watch deposit transaction
   const {
-    isLoading: isDepositConfirming,
+    isLoading: isApproving,
+    isSuccess: isApproveSuccess,
+    isError: isApproveError,
+  } = useWaitForTransaction({
+    hash: approveData?.hash,
+  });
+
+  const {
+    isLoading: isDepositing,
     isSuccess: isDepositSuccess,
-    error: depositError,
+    isError: isDepositError,
   } = useWaitForTransaction({
     hash: depositData?.hash,
-    onSuccess: () => {
-      handleDepositSuccess();
-    },
-    onError: (error) => handleTransactionError(error, "deposit", "onChain"),
   });
 
-  const needsApproval = allowance ? allowance < amountInWei : true;
-  const isLoading =
-    isApproveLoading ||
-    isApproveConfirming ||
-    isDepositLoading ||
-    isDepositConfirming;
+  // Enhanced approval transaction tracking
+  useEffect(() => {
+    if (isApproving) {
+      setTransactionState("approving");
+    } else if (isApproveSuccess) {
+      setTransactionState("approved");
+    } else if (isApproveError) {
+      setTransactionState("error");
+    }
+  }, [isApproving, isApproveSuccess, isApproveError]);
+
+  // Enhanced deposit transaction tracking
+  useEffect(() => {
+    if (isDepositing) {
+      setTransactionState("depositing");
+    } else if (isDepositSuccess) {
+      setTransactionState("success");
+      onSuccess?.();
+    } else if (isDepositError) {
+      setTransactionState("error");
+    }
+  }, [isDepositing, isDepositSuccess, isDepositError, onSuccess]);
 
   const handleApprove = async () => {
-    if (!approve) {
-      toast.error("Unable to initiate approval. Please try again.");
-      return;
+    try {
+      if (!lpToken.vault || amountInWei === BigInt(0)) {
+        throw new Error("Invalid vault address or amount");
+      }
+      approve?.({
+        args: [lpToken.vault, amountInWei],
+      });
+    } catch (error) {
+      console.error("Approval error:", error);
+      setTransactionState("error");
     }
-    const id = toast.info("Please confirm the approval in your wallet", {
-      autoClose: false,
-    });
-    setTransactionToastId(id);
-    approve({
-      args: [lpToken.vault, amountInWei],
-    });
   };
 
   const handleDeposit = async () => {
-    if (!deposit) {
-      toast.error("Unable to initiate deposit. Please try again.");
-      return;
+    try {
+      // Only allow deposit if approval is completed
+      if (transactionState !== "approved") {
+        return;
+      }
+
+      if (amountInWei === BigInt(0)) {
+        throw new Error("Invalid amount");
+      }
+
+      deposit?.({
+        args: [amountInWei],
+      });
+    } catch (error) {
+      console.error("Deposit error:", error);
+      setTransactionState("error");
     }
-    const id = toast.info("Please confirm the deposit in your wallet", {
-      autoClose: false,
-    });
-    setTransactionToastId(id);
-    deposit({
-      args: [amountInWei],
-    });
   };
 
-  const handleApproveSuccess = () => {
-    if (transactionToastId) {
-      toast.dismiss(transactionToastId);
-      setTransactionToastId(null);
-    }
-    toast.success(
-      <Stack spacing={1}>
-        <Typography variant="body1">Approval successful!</Typography>
-        <Typography variant="body2" sx={{ wordBreak: "break-all" }}>
-          Transaction: {approveData?.hash}
-        </Typography>
-      </Stack>
-    );
-  };
-
-  const handleDepositSuccess = () => {
-    if (transactionToastId) {
-      toast.dismiss(transactionToastId);
-      setTransactionToastId(null);
-    }
-    toast.success(
-      <Stack spacing={1}>
-        <Typography variant="body1">Deposit successful!</Typography>
-        <Typography variant="body2" sx={{ wordBreak: "break-all" }}>
-          Transaction: {depositData?.hash}
-        </Typography>
-      </Stack>
-    );
-    setTimeout(() => {
-      onSuccess?.();
-      onClose();
-    }, 1500);
-  };
-
-  const handleTransactionError = (
-    error: Error,
-    type: "approve" | "deposit",
-    stage: "pre" | "onChain"
-  ) => {
-    if (transactionToastId) {
-      toast.dismiss(transactionToastId);
-      setTransactionToastId(null);
-    }
-
-    const errorMessage = formatErrorMessage(
-      error?.message || "Transaction failed"
-    );
-    let userFriendlyMessage = "";
-
-    if (isUserRejection(errorMessage)) {
-      userFriendlyMessage = "Transaction rejected by user.";
-    } else if (isInsufficientFunds(errorMessage)) {
-      userFriendlyMessage = "Insufficient funds to complete the transaction.";
-    } else {
-      userFriendlyMessage = errorMessage;
-    }
-
-    toast.error(
-      <Stack spacing={1}>
-        <Typography variant="body1">
-          {`${type === "approve" ? "Approval" : "Deposit"} ${
-            stage === "pre" ? "Error" : "Failed"
-          }`}
-        </Typography>
-        <Typography variant="body2" color="error">
-          {userFriendlyMessage}
-        </Typography>
-      </Stack>
-    );
-  };
-
-  // Helper functions for error handling
-  const isUserRejection = (message: string): boolean => {
-    const rejectionPhrases = [
-      "user rejected",
-      "user denied",
-      "denied transaction",
-      "rejected by user",
-    ];
-    return rejectionPhrases.some((phrase) =>
-      message.toLowerCase().includes(phrase)
-    );
-  };
-
-  const isInsufficientFunds = (message: string): boolean => {
-    return message.toLowerCase().includes("insufficient funds");
-  };
-
-  const formatErrorMessage = (message: string): string => {
-    return message
-      .replace(/^Error: /i, "")
-      .replace(/\(action=.*\)/, "")
-      .replace(/\(code=.*\)/, "")
-      .trim();
-  };
-
-  const handleClose = () => {
-    if (transactionToastId) {
-      toast.dismiss(transactionToastId);
-      setTransactionToastId(null);
-    }
-    resetApprove?.();
-    resetDeposit?.();
-    onClose();
-  };
-
-  const truncateAddress = (address: Address) => {
+  const truncateAddress = (address: Address): string => {
     return `${address.slice(0, 6)}...${address.slice(-4)}`;
   };
 
+  const getStatusMessage = () => {
+    switch (transactionState) {
+      case "idle":
+        return "Please approve the transaction";
+      case "approving":
+        return "Approval in progress...";
+      case "approved":
+        return "Ready to deposit";
+      case "depositing":
+        return "Deposit in progress...";
+      case "success":
+        return "Transaction completed successfully";
+      case "error":
+        return "Transaction failed. Please try again.";
+      default:
+        return "Please approve the transaction";
+    }
+  };
+
+  // Determine if user can interact with modal
+  const canInteract = !isApproving && !isDepositing;
+
+  // Determine step states
+  const isApproveStep =
+    transactionState === "idle" || transactionState === "approving";
+  const isDepositStep = transactionState === "approved";
+
   return (
     <BaseModal
+      title="Auto Compounder"
       open={open}
-      onClose={handleClose}
-      title="Confirm Deposit"
-      aria-labelledby="deposit-confirmation-modal"
+      onClose={canInteract ? onClose : undefined}
+      aria-labelledby="auto-compounder-deposit-modal"
     >
-      <StepContainer>
-        <StepButton
-          $isActive={needsApproval && !isApproveSuccess}
-          disabled={!needsApproval || isApproveSuccess}
-        >
-          1. Approve
-        </StepButton>
-        <StepButton
-          $isActive={!needsApproval || isApproveSuccess}
-          disabled={needsApproval && !isApproveSuccess}
-        >
-          2. Submit
-        </StepButton>
-      </StepContainer>
-
-      <Typography
-        sx={{
-          fontFamily: "Inter",
-          fontSize: "16px",
-          color: "#FFFFFF",
-          letterSpacing: "0.01em",
-        }}
-      >
-        {needsApproval && !isApproveSuccess
-          ? "Please approve the transaction"
-          : "Please confirm your deposit"}
-      </Typography>
-
-      <Stack spacing={2}>
-        <Box>
-          <InputLabel>LP Address</InputLabel>
-          <InputField>
-            <Stack
-              gap={0.5}
-              direction="row"
-              justifyContent="space-between"
-              alignItems="center"
-            >
-              {truncateAddress(lpToken.poolAddress)}
-              <Tooltip
-                title={<Typography>{lpToken.poolAddress}</Typography>}
-                placement="top"
-              >
-                <IconButton size="small" sx={{ color: "text.secondary" }}>
-                  <RemoveRedEye fontSize="small" />
-                </IconButton>
-              </Tooltip>
-            </Stack>
-          </InputField>
-        </Box>
-
-        <Box>
-          <InputLabel>Spender Address</InputLabel>
-          <InputField>
-            <Stack
-              gap={0.5}
-              direction="row"
-              justifyContent="space-between"
-              alignItems="center"
-            >
-              {truncateAddress(lpToken.vault)}
-              <Tooltip
-                title={<Typography>{lpToken.vault}</Typography>}
-                placement="top"
-              >
-                <IconButton size="small" sx={{ color: "text.secondary" }}>
-                  <RemoveRedEye fontSize="small" />
-                </IconButton>
-              </Tooltip>
-            </Stack>
-          </InputField>
-        </Box>
-
-        <Box>
-          <InputLabel>Confirm quantity</InputLabel>
-          <InputField>
-            <Stack direction="row" alignItems="flex-end" gap={1}>
-              {amount}
-              <Typography color="text.secondary">{lpToken.name}</Typography>
-            </Stack>
-          </InputField>
-        </Box>
-      </Stack>
-
-      <Stack spacing={1.5}>
-        <ActionButton
-          variant="primary"
-          onClick={needsApproval ? handleApprove : handleDeposit}
-          disabled={
-            isLoading || // Disable during any loading state
-            (!needsApproval && !isApproveSuccess) || // Disable submit if approval not complete
-            (needsApproval && isApproveSuccess) // Disable approve if already approved
-          }
-        >
-          {isLoading ? (
-            <Stack direction="row" spacing={1} alignItems="center">
-              <CircularProgress size={20} color="inherit" />
-              <span>
-                {isApproveConfirming
-                  ? "CONFIRMING APPROVAL..."
-                  : isDepositConfirming
-                  ? "CONFIRMING DEPOSIT..."
-                  : "PREPARING..."}
-              </span>
-            </Stack>
-          ) : needsApproval ? (
-            "APPROVE"
-          ) : (
-            "DEPOSIT"
+      <ModalContent>
+        <ModalHeader>
+          <HeaderTitle>Confirm Deposit</HeaderTitle>
+          {canInteract && (
+            <CloseButton onClick={onClose} size="small">
+              <CloseIcon />
+            </CloseButton>
           )}
-        </ActionButton>
+        </ModalHeader>
 
-        <ActionButton
-          variant="secondary"
-          onClick={handleClose}
-          disabled={isApproveConfirming || isDepositConfirming}
-        >
-          GO BACK
-        </ActionButton>
-      </Stack>
+        <Box sx={{ p: "20px" }}>
+          <Stack spacing={3}>
+            <StepContainer>
+              <StepButton
+                $isActive={isApproveStep}
+                disabled={!isApproveStep || !canInteract}
+              >
+                1. {transactionState === "approved" ? "Approved" : "Approve"}
+              </StepButton>
+              <StepButton
+                $isActive={isDepositStep}
+                disabled={!isDepositStep || !canInteract}
+              >
+                2. Submit
+              </StepButton>
+            </StepContainer>
+
+            <Typography
+              sx={{
+                fontFamily: "Inter",
+                fontSize: "16px",
+                color: "#FFFFFF",
+                letterSpacing: "0.01em",
+              }}
+            >
+              {getStatusMessage()}
+            </Typography>
+
+            <Stack spacing={2}>
+              <Box>
+                <InputLabel>Contract address</InputLabel>
+                <InputField>{truncateAddress(lpToken.poolAddress)}</InputField>
+              </Box>
+
+              <Box>
+                <InputLabel>Spender Address</InputLabel>
+                <InputField>{truncateAddress(lpToken.vault)}</InputField>
+              </Box>
+
+              <Box>
+                <InputLabel>Confirm quantity</InputLabel>
+                <InputField>
+                  <Stack direction="row" alignItems="flex-end" gap={1}>
+                    {amount}
+                    <Typography color="text.secondary">
+                      {lpToken.name}
+                    </Typography>
+                  </Stack>
+                </InputField>
+              </Box>
+            </Stack>
+
+            <Stack spacing={1.5}>
+              {isApproveStep && (
+                <ActionButton
+                  variant="primary"
+                  onClick={handleApprove}
+                  disabled={!canInteract}
+                >
+                  {isApproving ? (
+                    <CircularProgress size={24} color="inherit" />
+                  ) : (
+                    "APPROVE"
+                  )}
+                </ActionButton>
+              )}
+
+              {isDepositStep && (
+                <ActionButton
+                  variant="primary"
+                  onClick={handleDeposit}
+                  disabled={!canInteract}
+                >
+                  {isDepositing ? (
+                    <CircularProgress size={24} color="inherit" />
+                  ) : (
+                    "DEPOSIT"
+                  )}
+                </ActionButton>
+              )}
+
+              <ActionButton
+                variant="secondary"
+                onClick={onClose}
+                disabled={!canInteract}
+              >
+                GO BACK
+              </ActionButton>
+            </Stack>
+          </Stack>
+        </Box>
+      </ModalContent>
     </BaseModal>
   );
 };
